🔨 Refactor: sign-up 수정
